# docs/swagger.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AlpaTrade API",
    "description": "AlpaTrade's production REST API for research, backtesting, validation, paper trading,\nreconciliation, reporting, and the primary LangChain DeepAgents assistant.\n\nUse `Authorization: Bearer <JWT>` for user clients. Trusted services may use\n`X-API-Key`; include `X-User-Id` only when acting on behalf of a specific user.\nAll order and autonomy endpoints are paper-only.",
    "contact": {
      "name": "Predictive Labs",
      "url": "https://alpatrade.chat/developers"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.36.0"
  },
  "servers": [
    {
      "url": "https://api.alpatrade.chat",
      "description": "Production"
    }
  ],
  "paths": {
    "/v2/leaderboard": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Leaderboard",
        "operationId": "leaderboard_v2_leaderboard_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 160,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "kind",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "",
                "live",
                "backtest"
              ],
              "type": "string",
              "default": "",
              "title": "Kind"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyList"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/v2/strategies": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Mine",
        "operationId": "mine_v2_strategies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyList"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ]
      },
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Create",
        "operationId": "create_v2_strategies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyId"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ]
      }
    },
    "/v2/strategies/{sid}": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Strategy",
        "operationId": "strategy_v2_strategies__sid__get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StrategyView"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "research"
        ],
        "summary": "Update",
        "operationId": "update_v2_strategies__sid__put",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "research"
        ],
        "summary": "Delete",
        "operationId": "delete_v2_strategies__sid__delete",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/strategies/{sid}/clone": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Clone",
        "operationId": "clone_v2_strategies__sid__clone_post",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CloneResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/strategies/{sid}/backtest": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Backtest",
        "operationId": "backtest_v2_strategies__sid__backtest_post",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "sid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Sid"
            }
          }
        ],
        "responses": {
          "202": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestJob"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/strategy-backtests/{job_id}": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Backtest Status",
        "operationId": "backtest_status_v2_strategy_backtests__job_id__get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BacktestStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/hedge-funds": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Funds",
        "operationId": "funds_v2_hedge_funds_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 160,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Period"
            }
          },
          {
            "name": "min_aum",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "",
                "100m",
                "1b",
                "10b",
                "100b",
                "lt100m"
              ],
              "type": "string",
              "default": "",
              "title": "Min Aum"
            }
          },
          {
            "name": "pos",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "",
                "concentrated",
                "focused",
                "diversified",
                "broad"
              ],
              "type": "string",
              "default": "",
              "title": "Pos"
            }
          },
          {
            "name": "rtype",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "holdings",
                "holdings_only",
                "combination",
                "notice",
                "any"
              ],
              "type": "string",
              "default": "holdings",
              "title": "Rtype"
            }
          },
          {
            "name": "holds",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 16,
              "default": "",
              "title": "Holds"
            }
          },
          {
            "name": "perf_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false,
              "title": "Perf Only"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "aum",
                "positions",
                "name",
                "filed",
                "ttm"
              ],
              "type": "string",
              "default": "aum",
              "title": "Sort"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundScreen"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/hedge-funds/performance": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Performance",
        "operationId": "performance_v2_hedge_funds_performance_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "method",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "quarter_end",
                "follow_filing"
              ],
              "type": "string",
              "default": "quarter_end",
              "title": "Method"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FundPerformance"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/hedge-funds/activists": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Activists",
        "operationId": "activists_v2_hedge_funds_activists_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "ticker",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 16,
              "default": "",
              "title": "Ticker"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/ipos": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Ipos",
        "operationId": "ipos_v2_ipos_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 300,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v2/ipos/pipeline": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "Pipeline",
        "operationId": "pipeline_v2_ipos_pipeline_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "ServiceApiKey": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BacktestJob": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          }
        },
        "type": "object",
        "required": [
          "job_id"
        ],
        "title": "BacktestJob"
      },
...
      "BacktestStatus": {
        "properties": {
          "state": {
            "type": "string",
            "enum": [
              "running",
              "done",
              "error"
            ],
            "title": "State"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "markdown": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Markdown"
          },
          "strategy_id": {
            "type": "integer",
            "title": "Strategy Id"
          }
        },
        "type": "object",
        "required": [
          "state",
          "strategy_id"
        ],
        "title": "BacktestStatus"
      },
...
      "CloneResult": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "created": {
            "type": "boolean",
            "title": "Created"
          }
        },
        "type": "object",
        "required": [
          "id",
          "created"
        ],
        "title": "CloneResult"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "StrategyId": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          }
        },
        "type": "object",
        "required": [
          "id"
        ],
        "title": "StrategyId"
      },
      "StrategyInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 160,
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "type": "string",
            "maxLength": 2000,
            "title": "Description",
            "default": ""
          },
          "skill_md": {
            "type": "string",
            "maxLength": 60000,
            "title": "Skill Md",
            "default": ""
          },
          "author_name": {
            "type": "string",
            "maxLength": 60,
            "title": "Author Name",
            "default": ""
          },
          "is_public": {
            "type": "boolean",
            "title": "Is Public",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "StrategyInput"
      },
      "StrategyList": {
        "properties": {
          "strategies": {
            "items": {
              "$ref": "#/components/schemas/StrategyView"
            },
            "type": "array",
            "title": "Strategies"
          },
          "total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "strategies"
        ],
        "title": "StrategyList"
      },
      "StrategyView": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "skill_md": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Skill Md"
          },
          "kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Kind"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Url"
          },
          "is_public": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Public"
          },
          "cloned_from_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloned From Id"
          },
          "author": {
            "type": "string",
            "title": "Author"
          },
          "metrics": {
            "additionalProperties": true,
            "type": "object",
            "title": "Metrics"
          },
          "methodology": {
            "type": "string",
            "title": "Methodology"
          },
          "rank": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rank"
          },
          "is_owner": {
            "type": "boolean",
            "title": "Is Owner",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "author",
          "metrics",
          "methodology"
        ],
        "title": "StrategyView"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "User JWT returned by POST /auth/login or /auth/register.",
        "scheme": "bearer"
      },
      "ServiceApiKey": {
        "type": "apiKey",
        "description": "Trusted service key. Add X-User-Id for tenant-scoped endpoints.",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  },
...
}